# zookeeper-rest-api/src/main/java/com/rfranco/zookeeperrestapi/autogenerated/dtos/UserSession.java
package com.rfranco.zookeeperrestapi.autogenerated.dtos;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.util.ArrayList;
import java.util.List;
import javax.validation.Valid;
import javax.validation.constraints.*;

/**
 * Information about the user session, including the user claims (permissions) and the session expiration.
 */
@ApiModel(description = "Information about the user session, including the user claims (permissions) and the session expiration.")
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-12-05T13:08:55.108+01:00")

public class UserSession   {
  @JsonProperty("expirationSeconds")
  private Integer expirationSeconds = null;

  @JsonProperty("claims")
  private List<String> claims = null;

  public UserSession expirationSeconds(Integer expirationSeconds) {
    this.expirationSeconds = expirationSeconds;
    return this;
  }

   /**
   * The number of seconds the session will take to expire.
   * @return expirationSeconds
  **/
  @ApiModelProperty(value = "The number of seconds the session will take to expire.")


  public Integer getExpirationSeconds() {
    return expirationSeconds;
  }

  public void setExpirationSeconds(Integer expirationSeconds) {
    this.expirationSeconds = expirationSeconds;
  }

  public UserSession claims(List<String> claims) {
    this.claims = claims;
    return this;
  }

  public UserSession addClaimsItem(String claimsItem) {
    if (this.claims == null) {
      this.claims = new ArrayList<String>();
    }
    this.claims.add(claimsItem);
    return this;
  }

   /**
   * The list of user claims.
   * @return claims
  **/
  @ApiModelProperty(value = "The list of user claims.")


  public List<String> getClaims() {
    return claims;
  }

  public void setClaims(List<String> claims) {
    this.claims = claims;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    UserSession userSession = (UserSession) o;
    return Objects.equals(this.expirationSeconds, userSession.expirationSeconds) &&
        Objects.equals(this.claims, userSession.claims);
  }

  @Override
  public int hashCode() {
    return Objects.hash(expirationSeconds, claims);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class UserSession {\n");
    
    sb.append("    expirationSeconds: ").append(toIndentedString(expirationSeconds)).append("\n");
    sb.append("    claims: ").append(toIndentedString(claims)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}